+++ test/unit/media.html5.js
@@ -92,3 +92,30 @@ test('should return a maybe for mp4 on OLD ANDROID', function() {
   vjs.IS_OLD_ANDROID = isOldAndroid;
   vjs.Html5.unpatchCanPlayType();
 });
+
+test('test playbackRate', function() {
+  var el, player, playbackRate, tech;
+
+  el = document.createElement('div');
+  el.innerHTML = '<div />';
+
+  player = {
+    id: function(){ return 'id'; },
+    el: function(){ return el; },
+    options_: {},
+    options: function(){ return {}; },
+    controls: function(){ return false; },
+    usingNativeControls: function(){ return false; },
+    on: function(){ return this; },
+    ready: function(){}
+  };
+
+  tech = new vjs.Html5(player, {});
+  tech.createEl();
+
+  tech.el_.playbackRate = 1.25;
+  strictEqual(tech.playbackRate(), 1.25);
+
+  tech['setPlaybackRate'](0.75);
+  strictEqual(tech.playbackRate(), 0.75);
+});