+++ CHANGELOG.md
@@ -18,6 +18,7 @@ CHANGELOG
 * Fixed compilation failures with LESS v1.7.0 and GRUNT v0.4.4 [[view](https://github.com/videojs/video.js/pull/1180)]
 * Added better error handling across the library [[view](https://github.com/videojs/video.js/pull/1197)]
 * Updated captions/subtiles file fetching to support cross-origin requests in older IE browsers [[view](https://github.com/videojs/video.js/pull/1095)]
+* Added support for playback rate switching [[view](https://github.com/videojs/video.js/pull/1132)]
 
 --------------------